requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
requests==2.31.0
python-dotenv>=0.19.0
ta==0.10.2
alpaca-trade-api>=2.0.0
yfinance>=0.1.63
ipykernel==6.25.0
scikit-learn>=0.24.0
joblib>=1.0.0
tqdm>=4.62.0 